created default package with less dependencies
