+++ src/microjson/__init__.py
@@ -1,6 +1,4 @@
 from .model import MicroJSON, GeoJSON
-from .utils import gather_example_files
-from .utils import OmeMicrojsonModel
-from .utils import MicrojsonBinaryModel
+
 
 __version__ = "0.1.11"
