+++ tests/test_original_schema.py
@@ -2,7 +2,7 @@ import json
 import jsonschema
 import pytest
 from jsonschema import validate
-from microjson.utils import gather_example_files
+from microjson.fileutils import gather_example_files
 
 # Load the original GeoJSON schema
 with open("external/GeoJSON.json") as f:
